requirements.txt
# DiceIQ Backend Dependencies
Flask==3.1.2
flask-cors==6.0.1
flask-compress==1.17
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Caching==2.3.1
Flask-Limiter==3.11.0
gunicorn==23.0.0
PyJWT==2.9.0
python-dotenv==1.1.1
Werkzeug==3.1.3
SQLAlchemy==2.0.43
stripe==11.3.0
pg8000
cloud-sql-python-connector[pg8000]
google-cloud-storage
google-cloud-secret-manager
